requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Authentication
django-allauth==0.60.0

# Stripe Payments
dj-stripe==2.8.3

# PDF Generation
weasyprint==60.1
django-weasyprint==2.3.0

# Database
psycopg[binary,pool]==3.1.14
dj-database-url==2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-results==2.5.1

# File Storage
boto3==1.34.14
django-storages==1.14.2
Pillow==10.2.0

# Environment & Security
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pandas==2.1.4

# Development
django-debug-toolbar==4.2.0
